+++ src/scenes.js
@@ -7,7 +7,7 @@
     written by Jens Mönig
     jens@moenig.org
 
-    Copyright (C) 2024 by Jens Mönig
+    Copyright (C) 2025 by Jens Mönig
 
     This file is part of Snap!.
 
@@ -53,7 +53,7 @@ normalizeCanvas, SnapSerializer, Costume, ThreadManager, IDE_Morph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.scenes = '2024-May-28';
+modules.scenes = '2025-October-27';
 
 // Projecct /////////////////////////////////////////////////////////
 
@@ -121,6 +121,8 @@ function Scene(aStageMorph) {
     this.unifiedPalette = false;
     this.showCategories = true;
     this.showPaletteButtons = true;
+    this.isTemplate = false;
+    this.template = null; // {name: str, version: str, hide: nested list}
 
     // cached IDE state
     this.sprites = new List();